chore: reducing code duplication
