+++ frontend/src/components/waste/DistrictVolumeListTable/index.tsx
@@ -1,12 +1,13 @@
 import { Column } from '@carbon/react';
-import { useEffect, useState, type FC } from 'react';
+
+import { useListTableState } from '../useListTableState';
 
 import { useDistrictVolumeListRowActions } from './actions';
 import { headers } from './constants';
 
 import type { PageableResponse } from '@/components/Form/TableResource/types';
 import type { DistrictVolumeListItem } from '@/services/districtvolumes.types';
-import type { SortDirectionType } from '@/services/types';
+import type { FC } from 'react';
 
 import TableResource from '@/components/Form/TableResource';
 import { useDistrictVolumeListQuery } from '@/config/react-query/hooks';
@@ -24,52 +25,11 @@ import './index.scss';
  * @returns The district volume list table view.
  */
 const DistrictVolumeListTable: FC = () => {
-  const [currentPage, setCurrentPage] = useState(0);
-  const [pageSize, setPageSize] = useState(10);
-  const [sort, setSort] = useState<Record<string, SortDirectionType>>({});
-  const [searchTrigger, setSearchTrigger] = useState(0);
-
   const getRowActions = useDistrictVolumeListRowActions();
 
-  const { data, isLoading, isFetching, isError, refetch } = useDistrictVolumeListQuery(
-    { page: currentPage, size: pageSize, sort },
-    { enabled: false, staleTime: Infinity },
-  );
-
-  /**
-   * Triggers a fetch with the given pagination state.
-   */
-  const executeSearch = (page: number, size: number) => {
-    setCurrentPage(page);
-    setPageSize(size);
-    setSearchTrigger((n) => n + 1);
-  };
-
-  useEffect(() => {
-    if (searchTrigger > 0) {
-      refetch();
-    }
-  }, [searchTrigger, refetch]);
-
-  useEffect(() => {
-    refetch();
-  }, [refetch]);
-
-  /**
-   * Applies a table page change and fetches the corresponding result set.
-   */
-  const handlePageChange = ({ page, pageSize }: { page: number; pageSize: number }) => {
-    const adjustedPage = Math.min(Math.max(page, 0), (data?.page.totalPages ?? 1) - 1);
-    executeSearch(adjustedPage, pageSize);
-  };
-
-  /**
-   * Applies updated sort keys and reruns the current search.
-   */
-  const handleSort = (sortingKeys: Record<string, SortDirectionType>) => {
-    setSort(sortingKeys);
-    executeSearch(currentPage, pageSize);
-  };
+  const { data, isLoading, isFetching, isError, handlePageChange, handleSort } = useListTableState({
+    queryHook: useDistrictVolumeListQuery,
+  });
 
   return (
     <Column lg={16} md={8} sm={4} className="configuration-column__content">
